# swagger.js
require('dotenv').config()
const swaggerAutogen = require('swagger-autogen')()
const resFormat = require('./lib/resFormat.js')
const port = process.env.PORT || '3000'

const doc = {
  info: {
    version: '1.0.0',
    title: 'My API',
    description: 'Documentation automatically generated by the <b>swagger-autogen</b> module.'
  },
  host: `localhost:${port}`,
  basePath: '/',
  schemes: ['http', 'https'],
  consumes: ['application/json'],
  produces: ['application/json'],
  tags: [
    {
      'name': 'User',
      'description': 'Endpoints'
    }
  ],
  securityDefinitions: {
    apiKeyAuth: {
      type: 'apiKey',
      in: 'header', // can be "header", "query" or "cookie"
      name: 'authorization', // name of the header, query parameter or cookie
      description: 'any description...'
    }
  },
  definitions: {
    resFormat: resFormat.return()
  }
}

const outputFile = './swagger-output.json'
const endpointsFiles = ['./routes/index.js']

swaggerAutogen(outputFile, endpointsFiles, doc).then(() => {
  require('./server.js') // Your project's root file
})